% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{applyInteractionType}
\alias{applyInteractionType}
\title{generate pairs of interactions according to interaction types}
\usage{
applyInteractionType(I, pair, interType)
}
\arguments{
\item{I}{Matrix: defining the interaction between each pair of species}

\item{pair}{Numeric: a vector with a length of 2, indicating the 2 focusing
species in the process of applying the interaction types}

\item{interType}{Character: one of 'mutualism', 'commensalism', 'parasitism',
'amensalism', or 'competition'. Defining the interaction type}
}
\value{
A matrix of interaction types with one pair changed
}
\description{
A helper function to be used in combination with
getInteractions()
}
\examples{
applyInteractionType(matrix(0, nrow = 3, ncol = 3), 
                     pair = c(1,2), 
                     interType = "parasitism")
}
