% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertToSE.R
\name{convertToSE}
\alias{convertToSE}
\title{It helps to store any matrix in \linkS4class{SummarizedExperiment}
object class data container(\code{SE}). The resulting abundance matrix from the
simulation functions used in \code{miaSim} can be used.}
\usage{
convertToSE(matrix, tse = FALSE, ...)
}
\arguments{
\item{matrix}{: A matrix-like or list of matrix-like objects.
Rows: genes, genomic coordinates, etc. Columns: samples, cells, etc.}

\item{tse}{: Logical, decides if the object is stored in
\linkS4class{TreeSummarizedExperiment} or \linkS4class{SummarizedExperiment}}

\item{...}{: additional objects that can be implemented in the \code{SE} object.}
}
\value{
\linkS4class{SummarizedExperiment} an object containing abundance
matrix
}
\description{
It helps to store any matrix in \linkS4class{SummarizedExperiment}
object class data container(\code{SE}). The resulting abundance matrix from the
simulation functions used in \code{miaSim} can be used.
}
\examples{
ExampleHubbellRates <- miaSim::simulateHubbellRates(
    community_initial = c(0,5,10), migration_p = 0.01,
    metacommunity_p = NULL, k_events = 1, growth_rates = NULL, norm = FALSE,
    t.end=1000)

HubbellSE <- convertToSE(matrix = ExampleHubbellRates$abundancematrix,
                         colData = ExampleHubbellRates$time,
                         metadata = ExampleHubbellRates$metadata)

}
