% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertToSE.R
\name{convertToTreeSE}
\alias{convertToTreeSE}
\title{Convert simulated microbiome data to \linkS4class{TreeSummarizedExperiment}(\code{TreeSE}) data container}
\usage{
convertToTreeSE(assay, output, ...)
}
\arguments{
\item{assay}{is a matrix-like or list of matrix-like object.
Rows refer to taxa and columns refer to samples.}

\item{output}{character value for storing the matrix in
\linkS4class{TreeSummarizedExperiment} (\code{output = TreeSE}).}

\item{...}{: additional parameters to pass}
}
\value{
\linkS4class{TreeSummarizedExperiment} data object containing abundance matrix
}
\description{
This function converts simulated microbial community data to
\linkS4class{TreeSummarizedExperiment}(\code{TreeSummarizedExperiment}) format.
}
\details{
The abundance matrix from the simulation functions
in \code{miaSim} can be converted to \linkS4class{TreeSummarizedExperiment}
class object.

Storing the data in \linkS4class{TreeSummarizedExperiment} enables access to
various Bioconductor packages and tools that extend the
this class. This offers data and metadata
synchronization, while accommodating specialized data structures for
particular scientific applications.

Examples for \code{TreeSE} object manipulation and analysis can be found at
\url{https://microbiome.github.io/OMA}
}
\examples{
# Simulate time series data
x <- simulateHubbellRates(
    community_initial = c(0,5,10), migration_p = 0.01,
    metacommunity_p = NULL, k_events = 1, growth_rates = NULL, norm = FALSE,
    t_end=1000)

# Convert into TreeSE format.
tse <- TreeSummarizedExperiment(
                         assays = list(counts = x$counts), 
                         colData = x$time,
                         metadata = x$metadata)

}
