% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{createParamList}
\alias{createParamList}
\title{Create a list of parameter.}
\usage{
createParamList(input_param, n_repeat, replace_by_zero = FALSE)
}
\arguments{
\item{input_param}{Vector: the input parameter for a simulation.}

\item{n_repeat}{Integer: the number to repeat the parameter.}

\item{replace_by_zero}{Boolean: whether to replace certain elements in result with
0 using internal function \code{.replaceByZero}.}
}
\value{
A list of parameters which is ready as input of param_iter for
\code{generateSimulations}.
}
\description{
This function is intended to generate a list of parameters such as initial
community, to prepare the parameter param_iter in \code{generateSimulations}.
}
\examples{
paramx0 <- createParamList(input_param = rep(99, 10), n_repeat = 10, replace_by_zero = TRUE)
}
