% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.isPosInt}
\alias{.isPosInt}
\title{Check whether a number is a positive integer}
\usage{
.isPosInt(x, tol = .Machine$double.eps^0.5)
}
\arguments{
\item{x}{Numeric number to test}

\item{tol}{Numeric tolerance of detection}
}
\value{
A logical value: whether the number is a positive integer.
}
\description{
Check whether a number is a positive integer
}
