% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.rdirichlet}
\alias{.rdirichlet}
\title{Generate dirichlet random deviates}
\usage{
.rdirichlet(n, alpha)
}
\arguments{
\item{n}{Number of random vectors to generate.}

\item{alpha}{Vector containing shape parameters.}
}
\value{
a vector containing the Dirichlet density
}
\description{
Generate dirichlet random deviates
}
\examples{
dirichletExample <- .rdirichlet(1, c(1, 2, 3))

}
