% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getCommunity}
\alias{getCommunity}
\title{Get the community from one simulation}
\usage{
getCommunity(simulation, t_end = NULL)
}
\arguments{
\item{simulation}{A list generated from simulateXXX function}

\item{t_end}{Numeric: end time of the simulation. If NULL, the last t_end in
the simulation will be used to extract the final state.
(default: \code{t_end = NULL})}
}
\value{
a subset of dataframe simulation$matrix, containing the composition
at t_end.
}
\description{
Get the community composition at specific time point in a simulation
(default: last time point)
}
\examples{
# single use: extract community composition at t_end = 200 from a simulation
getCommunity(simulateGLV(n_species = 3, t_end = 1000), t_end = 200)
# apply this function to previously generated list of simulations
GLVSimus <- generateSimulations("simulateGLV", list(n_species = 3, t_end = 1000), n_instances = 5)
GLVSimusCom <- as.data.frame(do.call(rbind, lapply(GLVSimus, getCommunity)))

}
