% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getInteractions}
\alias{getInteractions}
\title{generate interactions according to five types of interactions and
their weights}
\usage{
getInteractions(n_species, weights, connectance)
}
\arguments{
\item{n_species}{Integer: defining the dimension of matrix of interaction}

\item{weights}{Numeric: defining the weights of mutualism, commensalism,
parasitism, amensalism, and competition in all interspecies interactions.}

\item{connectance}{Numeric: defining the density of the interaction network.
Ranging from 0 to 1}
}
\value{
A matrix of interactions with all interactions changed according to
the weights and connectance.
}
\description{
generate interactions according to five types of interactions and
their weights
}
\examples{
I <- getInteractions(n_species = 10, 
                     weights = c(1,1,1,1,1), 
                     connectance = 0.5)
}
