% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getRowMax}
\alias{getRowMax}
\title{only get the maximum value from a row}
\usage{
getRowMax(aRow)
}
\arguments{
\item{aRow}{Vector: a row in a matrix/dataframe}
}
\value{
A vector of numbers with only the first maximun value kept, and other
numbers changed to 0
}
\description{
A helper function to be used in combination with apply()
}
