% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HubbellSE.R
\name{HubbellSE}
\alias{HubbellSE}
\title{Hubbell community simulation}
\usage{
HubbellSE(N, M, I, d, m, tskip = 0, tend, norm = FALSE)
}
\arguments{
\item{N}{Integer scalar indicating the amount of different species initially
in the local community}

\item{M}{Integer scalar indicating the amount of different species in the
metacommunity,including those of the local community}

\item{I}{Integer scalar indicating the fixed amount of individuals in the
local community (default: \code{I = 1000})}

\item{d}{Integer scalar indicating fixed amount of deaths of local community
individuals in each generation (default: \code{d = 10})}

\item{m}{Numeric scalar indicating immigration rate: the probability that a
death in the local community is replaced by a migrant of the metacommunity
rather than by the birth of a local community member
(default: \code{m = 0.02})}

\item{tskip}{Integer scalar indicating the number of generations that should
not be included in the outputted species abundance matrix.
(default: \code{tskip = 0})}

\item{tend}{Integer scalar indicating the number of simulations
to be simulated}

\item{norm}{Logical scalar to indicate whether the time series should
be returned with the abundances as proportions (norm = TRUE) or the raw
counts (norm = FALSE, default)}
}
\value{
\code{HubbellSE} returns a \code{\link{SummarizedExperiment}} object
}
\description{
Neutral species abundances simulation turned into
\linkS4class{SummarizedExperiment} according to the Hubbell model.
}
\examples{
HubbellSE(N = 8, M = 10, I = 1000, d = 50, m = 0.02, tend = 100)
}
\references{
Rosindell, James et al. "The unified neutral theory of
biodiversity and biogeography at age ten." Trends in ecology & evolution
vol. 26,7 (2011).
}
