% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{isPositiveInteger}
\alias{isPositiveInteger}
\title{check whether a number is a positive integer}
\usage{
isPositiveInteger(x, tol = .Machine$double.eps^0.5)
}
\arguments{
\item{x}{Numeric number to test}

\item{tol}{Numeric tolerance of detection}
}
\value{
A logical value: whether the number is a positive integer.
}
\description{
check whether a number is a positive integer
}
