% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{isZeroOrPositiveInteger}
\alias{isZeroOrPositiveInteger}
\title{check whether a number is zero or positive integer}
\usage{
isZeroOrPositiveInteger(x, tol = .Machine$double.eps^0.5)
}
\arguments{
\item{x}{Numeric number to test}

\item{tol}{Numeric tolerance of detection}
}
\value{
A logical value: whether the number is zero or positive integer.
}
\description{
check whether a number is zero or positive integer
}
