% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{makeHeatmap}
\alias{makeHeatmap}
\title{simple heatmap plot of matrix}
\usage{
makeHeatmap(
  matrix.A,
  title = "Consumption/production matrix",
  y.label = "resources",
  x.label = "species",
  midpoint_color = NULL,
  lowColor = "red",
  midColor = "white",
  highColor = "blue"
)
}
\arguments{
\item{matrix.A}{a matrix. Can be a matrix generated with randomA or randomE}

\item{title, x.label, y.label, midpoint_color, lowColor, midColor, highColor}{detailed controls of the heatmap}
}
\value{
a ggplot type of plot.
}
\description{
simple heatmap plot of matrix
}
\examples{
dense_A <- randomA(n_species = 10, 
    scale_off_diagonal = 1, 
    diagonal = -1.0, 
    connectance = 0.9)
makeHeatmap(dense_A, lowColor = 'blue', highColor = 'red')
}
