% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{makePlot}
\alias{makePlot}
\title{simple line plot of species}
\usage{
makePlot(
  out_matrix,
  title = "abundance of species by time",
  obj = "species",
  y.label = "x.t"
)
}
\arguments{
\item{out_matrix}{a result matrix. Rows are timepoints and columns are
species. Last column named "time" indicates the timepoint of each row in the
simulation.}

\item{title, obj, y.label}{title, x, and y axis of the plot}
}
\value{
a ggplot type of plot.
}
\description{
simple line plot of species
}
\examples{
ExampleCR <- simulateConsumerResource(n_species = 2, 
    n_resources = 4)
makePlot(ExampleCR$matrix)
}
