% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{makePlotRes}
\alias{makePlotRes}
\title{simple line plot of resources}
\usage{
makePlotRes(out_matrix, title = "quantity of compounds by time")
}
\arguments{
\item{out_matrix}{a result matrix of resources. Rows are timepoints and
columns are resources. Last column named "time" indicates the timepoint of
each row in the simulation.}

\item{title}{title of the plot}
}
\value{
a ggplot type of plot.

a ggplot type of plot.
}
\description{
simple line plot of resources
}
\examples{
ExampleCR <- simulateConsumerResource(n_species = 2, 
    n_resources = 4)
makePlotRes(ExampleCR$resources)

}
