% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomA.R
\name{randomA}
\alias{randomA}
\title{Generate random uniform interaction matrix}
\usage{
randomA(
  n_species,
  d = -0.5,
  min_strength = -0.5,
  max_strength = 0.5,
  connectance = 0.02,
  symmetric = FALSE
)
}
\arguments{
\item{n_species}{integer number of species}

\item{d}{numeric diagonal values (should be negative)
(default: \code{d = -0.5})}

\item{min_strength}{numeric value of minimal off-diagonal interaction
strength (default: \code{min_strength = -0.5})}

\item{max_strength}{numeric value of maximal off-diagonal interaction
strength (default: \code{max_strength = 0.5})}

\item{connectance}{numeric interaction probability
(default: \code{connectance = 0.02})}

\item{symmetric}{logical scalar returning a symmetric interaction matrix
(default: \code{symmetric=FALSE})}
}
\value{
\code{randomA} returns a matrix A with dimensions (n_species x n_species)
}
\description{
Generate random simplified interaction matrix from a uniform distribution.
}
\examples{
high_inter_A <- randomA(n_species = 10, d = -0.4, min_strength = -0.8,
                                    max_strength = 0.8, connectance = 0.5)

low_inter_A <- randomA(n_species = 10, connectance = 0.01)

}
