% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomE.R
\name{randomE}
\alias{randomE}
\title{Generate random efficiency matrix}
\usage{
randomE(
  n.species,
  n.resources,
  min.con = round(n.resources/4),
  max.con = round(n.resources/3),
  min.prod = round(n.resources/6),
  max.prod = round(n.resources/4),
  maintenance = 0.5
)
}
\arguments{
\item{n.species}{integer number of species}

\item{n.resources}{integer number of resources}

\item{min.con}{integer minimum number of resources consumed by each species}

\item{max.con}{integer maximum number of resources consumed by each species}

\item{min.prod}{integer minimum number of resources produced by each species}

\item{max.prod}{integer maximum number of resources produced by each species}

\item{maintenance}{numeric value between 0~1 the proportion of resources used
to maintain the living of microorganisms. 0 means all the resources will be
used for the reproduction of microorganisms, and 1 means all the resources
would be used to maintain the living of organisms and no resources would be
left for their growth(reproduction).}
}
\value{
\code{randomE} returns a matrix E with dimensions (n.species x n.resources),
and each row represents a species.
}
\description{
Generate random efficiency matrix for consumer resource model from Dirichlet
distribution. Positive efficiencies indicate the consumption of resources,
whilst negatives indicate that the species would produce the resource.
}
\examples{
# example with specific parameters
ExampleEfficiencyMatrix <- randomE(n.species = 3, n.resources = 6,
min.con = 3, max.con = 4, min.prod = 1, max.prod = 1, maintenance = 0.4)
# example with minimum parameters
ExampleEfficiencyMatrix2 <- randomE(n.species = 5, n.resources = 12)

}
