% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateConsumerResource.R
\name{simulateConsumerResource}
\alias{simulateConsumerResource}
\title{Consumer-resource model simulation}
\usage{
simulateConsumerResource(
  n_species,
  n_resources,
  eff = randomE(n_species, n_resources),
  consumers = runif(n = n_species, min = 0.1, max = 10),
  resources = runif(n = n_resources, min = 1, max = 100),
  mumax = rep(1, n_species),
  k_table = matrix(rgamma(n = n_species * n_resources, shape = 50, rate = 0.25), nrow =
    n_species),
  t_end = 1000,
  ...
)
}
\arguments{
\item{n_species}{integer number of species}

\item{n_resources}{interger number of resources}

\item{eff}{a matrix of efficiency. How efficient are resources
converted into biomass, negative values represent excreted resources
(default: \code{eff = randomE(n_species, n_resources)})}

\item{consumers}{numeric vector of species
(default: \code{consumers = runif(n = n_species, min = 0.1, max = 10)})}

\item{resources}{numeric vector of resources
(default: \code{resources = runif(n = n_resources, min = 1, max = 100)})}

\item{mumax}{numeric vector of maximum mu of species
(default: \code{mumax = rep(1, n_species)})}

\item{k_table}{a matrix of K values in monod model
(default: \code{k_table = matrix(rgamma(n=n_species*n_resources,
shape = 50, rate = 0.25), nrow = n_species)})}

\item{t_end}{numeric scalar indicating the final time of the simulation
(default: \code{t_end = 1000})}

\item{...}{additional parameters including 't_start', 't_step', and 't_store'}
}
\value{
an abundance matrix with species and resources abundance as rows and
time points as columns
}
\description{
Simulates a community time series using the consumer-resource model.
The change rate of each species was defined as
\code{dx/dt = mumax*sum(monod)*X}, where
mumax is the vector of maximum growth rates for the species,
monod is the monod growth rate, S/(Ks+S), where S is the concentration of the
limiting resource, and Ks is the half-velocity constant for species X and S.
X is the vector of abundances of species.
The concentrations of resource will be set to 0 if they were calculated
less than 0 during the iteration.
}
\examples{
# example1 users provide least parameters.
ExampleConsumerResource <- simulateConsumerResource(n_species = 2,
n_resources = 4)

}
\seealso{
\code{\link[miaSim:convertToSE]{convertToSE}}
}
