% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{simulateEventTimes}
\alias{simulateEventTimes}
\title{Generate a vector of event times}
\usage{
simulateEventTimes(t_events = NULL, t_duration = NULL, t_end = 1000, ...)
}
\arguments{
\item{t_events}{Numeric vector; starting time of the events}

\item{t_duration}{Numeric vector; duration of the events}

\item{t_end}{Numeric: end time of the simulation}

\item{...}{: additional parameters to pass to simulationTimes, including
t_start, t_step, and t_store.}
}
\value{
A vector of time points in the simulation
}
\description{
Generate a vector of event times
}
\examples{
tEvent <- simulateEventTimes(
    t_events = c(10, 50, 100),
    t_duration = c(1, 2, 3),
    t_end = 100,
    t_store = 100,
    t_step = 1
)
}
