% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{TimeSeriesSE}
\alias{TimeSeriesSE}
\title{Simulate time series with the generalized Lotka-Volterra model and turn into SummarizedExperiment object}
\arguments{
\item{N}{species number}

\item{A}{interaction matrix}

\item{b}{growth rates}

\item{x}{initial abundances}

\item{tend}{timepoints}

\item{norm}{return normalised abundances (proportions in each generation)}
}
\value{
a SummarizedExperiment object that has abundance matrix in assay,
colData that includes sampleID and time, rowData that includes species name, ASV levels,
and the taxonomic information:
\itemize{
\item{Kingdom}
\item{Phylum}
\item{Class}
}
\description{
Simulate a community time series using the generalized Lotka-Volterra model,
defined as dx/dt = x(b+Ax), where x is the vector of species abundances,
A is the interaction matrix and growth_rates the vector of growth rates.

The abundance matrix that is obtained from the generalized Lotka-Volterra model, is used to construct
\code{\link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment}} object
}
\examples{
result <- glv(N = 4, A = powerlawA(n = 4, alpha = 2), tend = 1000)
SE <- conversionSE(result)
}
