% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertToSE.R
\name{convertToSE}
\alias{convertToSE}
\title{\linkS4class{SummarizedExperiment}(\code{SE}) or
\code{TreeSE} construction function}
\usage{
convertToSE(assay, output, ...)
}
\arguments{
\item{assay}{is a matrix-like or list of matrix-like object.
Rows refer to taxa and columns refer to samples.}

\item{output}{character value for storing the matrix in either
\linkS4class{TreeSummarizedExperiment} (\code{output = TSE}) or
\linkS4class{SummarizedExperiment} (default: \code{output = SE})}

\item{...}{: additional parameters to pass for
the \code{SE} object.}
}
\value{
\linkS4class{SummarizedExperiment} an object containing abundance
matrix
}
\description{
The abundance matrix from the simulation functions
in \code{miaSim} can be converted to \linkS4class{SummarizedExperiment}
class object.
}
\details{
Storing the data in \linkS4class{SummarizedExperiment} enables access to
various Bioconductor packages and tools that extend the
\linkS4class{SummarizedExperiment} class.
This offers data and metadata
synchronization, while accommodating specialized data structures for
particular scientific applications.

Further examples for \code{SE} object manipulation and analysis can be found at
\url{https://microbiome.github.io/OMA}
}
\examples{
x <- simulateHubbellRates(
    community_initial = c(0,5,10), migration_p = 0.01,
    metacommunity_p = NULL, k_events = 1, growth_rates = NULL, norm = FALSE,
    t_end=1000)

<<<<<<< HEAD
HubbellSE <- convertToSE(matrix = x$counts,
                         colData = x$time,
                         metadata = x$metadata)
=======
HubbellSE <- convertToSE(assay = ExampleHubbellRates$counts,
                        colData = ExampleHubbellRates$time,
                        metadata = ExampleHubbellRates$metadata)
>>>>>>> main

}
