% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertToSE.R
\name{convertToSE}
\alias{convertToSE}
\title{Old conversion function (to be deprecated)}
\usage{
convertToSE(assay, output, ...)
}
\arguments{
\item{assay}{is a matrix-like or list of matrix-like object.
Rows refer to taxa and columns refer to samples.}

\item{output}{character value for storing the matrix in
\linkS4class{TreeSummarizedExperiment} (\code{output = TreeSE}).}

\item{...}{: additional parameters to pass}
}
\value{
SummarizedExperiment data object containing abundance matrix
}
\description{
Old conversion function (to be deprecated)
}
