% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomA.R
\docType{methods}
\name{randomA}
\alias{randomA}
\alias{randomA-numeric}
\alias{randomA,numeric-method}
\title{Generate random uniform interaction matrix}
\usage{
randomA(
  n.species,
  d = -0.5,
  min.strength = -0.5,
  max.strength = 0.5,
  connectance = 0.02
)
}
\arguments{
\item{n.species}{integer number of species}

\item{d}{numeric diagonal values (should be negative)
(default: \code{d = -0.5})}

\item{min.strength}{numeric minimal off-diagonal interaction strength
(default: \code{min.strength = -0.5})}

\item{max.strength}{numeric maximal off-diagonal interaction strength
(default: \code{max.strength = 0.5})}

\item{connectance}{numeric (interaction probability)
(default: \code{connectance = 0.02})}
}
\value{
\code{randomA} returns a matrix A with n.species times rows and
n.species times columns
}
\description{
Generate random simplified interaction matrix from a uniform distribution.
}
\examples{
high_inter_A <- randomA(10, d = -0.4, min.strength = -0.8,
                                    max.strength = 0.8, connectance = 0.5)
low_inter_A <- randomA(10, c = 0.01)

}
