% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomA.R
\name{randomA}
\alias{randomA}
\title{Generate random uniform interaction matrix}
\usage{
randomA(
  n.species,
  d = -0.5,
  min.strength = -0.5,
  max.strength = 0.5,
  connectance = 0.02,
  symmetric = FALSE
)
}
\arguments{
\item{n.species}{integer number of species}

\item{d}{numeric diagonal values
(default: \code{d = -0.5})}

\item{min.strength}{numeric minimal off-diagonal interaction strength
(default: \code{min.strength = -0.5})}

\item{max.strength}{numeric maximal off-diagonal interaction strength
(default: \code{max.strength = 0.5})}

\item{connectance}{numeric (interaction probability)
(default: \code{connectance = 0.02})}

\item{symmetric}{logical return a symmetric interaction matrix
(default: \code{symmetric=FALSE})}
}
\value{
\code{randomA} returns a matrix A with dimensions (n.species x n.species)
}
\description{
Generate random simplified interaction matrix from a uniform distribution.
}
\examples{
high_inter_A <- randomA(n.species = 10, d = -0.4, min.strength = -0.8,
                                    max.strength = 0.8, connectance = 0.5)

low_inter_A <- randomA(n.species = 10, connectance = 0.01)

}
