% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomE.R
\name{randomE}
\alias{randomE}
\title{Generate random efficiency matrix}
\usage{
randomE(
  n_species,
  n_resources,
  min_con = round(n_resources/4),
  max_con = round(n_resources/3),
  min_prod = round(n_resources/6),
  max_prod = round(n_resources/4),
  maintenance = 0.5
)
}
\arguments{
\item{n_species}{integer number of species}

\item{n_resources}{integer number of resources}

\item{min_con}{integer minimum number of resources consumed by each species}

\item{max_con}{integer maximum number of resources consumed by each species}

\item{min_prod}{integer minimum number of resources produced by each species}

\item{max_prod}{integer maximum number of resources produced by each species}

\item{maintenance}{numeric value between 0~1 the proportion of resources used
to maintain the living of microorganisms. 0 means all the resources will be
used for the reproduction of microorganisms, and 1 means all the resources
would be used to maintain the living of organisms and no resources would be
left for their growth(reproduction).}
}
\value{
\code{randomE} returns a matrix E with dimensions (n_species x n_resources),
and each row represents a species.
}
\description{
Generate random efficiency matrix for consumer resource model from Dirichlet
distribution. Positive efficiencies indicate the consumption of resources,
whilst negatives indicate that the species would produce the resource.
}
\examples{
# example with minimum parameters
ExampleEfficiencyMatrix2 <- randomE(n_species = 5, n_resources = 12)

}
