% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomE.R
\name{randomE}
\alias{randomE}
\title{Generate random efficiency matrix}
\usage{
randomE(
  n.species,
  n.resources,
  min.con = round(n.resources/4),
  max.con = round(n.resources/3),
  min.prod = round(n.resources/6),
  max.prod = round(n.resources/4),
  sd = 0.5
)
}
\arguments{
\item{n.species}{Integer: number of species}

\item{n.resources}{Integer: number of resources}

\item{min.con}{Integer: minimum number of resources consumed by each species}

\item{max.con}{Integer: maximum number of resources consumed by each species}

\item{min.prod}{Integer: minimum number of resources produced by each species}

\item{max.prod}{Integer: maximum number of resources produced by each species}

\item{sd}{Numeric: standard deviation of the normal distribution}
}
\value{
\code{randomE} returns a matrix E with dimensions (n.species x n.resources),
and each row represents a species.
}
\description{
Generate random efficiency matrix for consumer resource model from a normal
distribution. Positive efficiencies indicate the consumption of resources,
whilst negatives indicate that the species would produce the resource.
Efficiencies larger than 1 would be replaced by 1.
}
\examples{
# example with specific parameters
ExampleEfficiencyMatrix <- randomE(n.species = 3, n.resources = 6,
min.con = 3, max.con = 4, min.prod = 1, max.prod = 1, sd = 0.4)
# example with minimum parameters
ExampleEfficiencyMatrix2 <- randomE(n.species = 5, n.resources = 12)

}
