% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateGLV.R
\docType{methods}
\name{simulateGLV}
\alias{simulateGLV}
\alias{simulateGLV-numeric}
\alias{simulateGLV,numeric-method}
\title{Generalized Lotka-Volterra (gLV) simulation}
\usage{
simulateGLV(
  n.species,
  A,
  tend = 1000,
  x = runif(n.species),
  b = runif(n.species),
  norm = FALSE
)
}
\arguments{
\item{n.species}{Integer: number of species}

\item{A}{interaction matrix}

\item{tend}{Integer: timepoints (default: \code{tend = 1000})}

\item{x}{Numeric: initial abundances}

\item{b}{Numeric: growth rates}

\item{norm}{Logical scalar: returns normalised abundances (proportions
in each generation) (default: \code{norm = FALSE})}
}
\value{
\code{simulateGLV} returns a \linkS4class{SummarizedExperiment} object
containing abundance matrix
}
\description{
Simulates time series with the generalized Lotka-Volterra model and forms a
\linkS4class{SummarizedExperiment} object.
}
\details{
Simulates a community time series using the generalized Lotka-Volterra model,
defined as dx/dt = x(b+Ax), where x is the vector of species abundances,
A is the interaction matrix and b is the vector of growth rates.

The resulting abundance matrix model is used to construct
\linkS4class{SummarizedExperiment} object.
}
\examples{
row_data <- data.frame(Kingdom = "Animalia",
                Phylum = rep(c("Chordata", "Echinodermata"), c(500, 500)),
                Class = rep(c("Mammalia", "Asteroidea"), each = 500),
                ASV = paste0("X", seq_len(1000)),
                row.names = rownames(paste0("species", seq_len(1000))),
                stringsAsFactors = FALSE)

row_data <- t(row_data)

col_data <- DataFrame(sampleID = seq_len(1000),
                    time = as.Date(1000, origin = "2000-01-01"),
                    row.names = colnames(paste0("sample", seq_len(1000))))

A <- miaSim::powerlawA(4, alpha = 4)

SEobject <- simulateGLV(n.species = 4, A, tend = 1000)
rowData(SEobject) <- row_data
colData(SEobject) <- col_data

}
