% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateGLV.R
\name{simulateGLV}
\alias{simulateGLV}
\title{Generalized Lotka-Volterra (gLV) simulation}
\usage{
simulateGLV(
  n.species,
  A,
  x = runif(n.species),
  b = runif(n.species),
  sigma.drift = 0.01,
  sigma.epoch = 0.3,
  sigma.external = 0.3,
  p.epoch = 0.01,
  t.external_events = c(12, 36, 48),
  t.external_durations = c(3, 10, 99),
  stochastic = FALSE,
  norm = FALSE,
  t.end = 1000,
  ...
)
}
\arguments{
\item{n.species}{Integer: number of species}

\item{A}{interaction matrix}

\item{x}{Numeric: initial abundances}

\item{b}{Numeric: growth rates}

\item{sigma.drift}{Numeric: degree of drift (turnover of species) in each
time step.
(default: \code{sigma.drift = 0.01})}

\item{sigma.epoch}{Numeric: degree of epoch change of community
(default: \code{sigma.epoch = 0.3})}

\item{sigma.external}{Numeric: degree of external events/disturbances
(default: \code{sigma.external = 0.3})}

\item{p.epoch}{Numeric: probability/frequency of inherit periodic changes of
community
(default: \code{p.epoch = 0.01})}

\item{t.external_events}{Numeric: starting times of external events
(default: \code{t.external_events = c(12, 36, 48)})}

\item{t.external_durations}{Numeric: durations of external events
(default: \code{t.external_durations = c(3, 10, 99)})}

\item{stochastic}{Logical: whether the gLV model should be stochastic
(default: \code{stochastic = FALSE})}

\item{norm}{Logical scalar: returns normalised abundances (proportions
in each generation) (default: \code{norm = FALSE})}

\item{t.end}{Numeric: simulation end time (default: \code{t.end = 1000})}

\item{...}{additional parameters including 't.start', 't.step', and 't.store'}
}
\value{
\code{simulateGLV} returns a \linkS4class{SummarizedExperiment} object
containing abundance matrix
}
\description{
Simulates time series with the generalized Lotka-Volterra model and forms a
\linkS4class{SummarizedExperiment} object.
}
\details{
Simulates a community time series using the generalized Lotka-Volterra model,
defined as dx/dt = diag(x)(b+Ax), where x is the vector of species abundances
,diag(x) is a diagonal matrix with the diagonal values set to x.
A is the interaction matrix and b is the vector of growth rates.

The resulting abundance matrix model is used to construct
\linkS4class{SummarizedExperiment} object.
}
\examples{
row_data <- data.frame(Kingdom = "Animalia",
                Phylum = rep(c("Chordata", "Echinodermata"), c(500, 500)),
                Class = rep(c("Mammalia", "Asteroidea"), each = 500),
                ASV = paste0("X", seq_len(1000)),
                row.names = rownames(paste0("species", seq_len(1000))),
                stringsAsFactors = FALSE)

row_data <- t(row_data)

col_data <- DataFrame(sampleID = seq_len(1000),
                    time = as.Date(1000, origin = "2000-01-01"),
                    row.names = colnames(paste0("sample", seq_len(1000))))

A <- miaSim::powerlawA(4, alpha = 1.01)

SEobject <- simulateGLV(n.species = 4, A, t.end = 1000)
rowData(SEobject) <- row_data
colData(SEobject) <- col_data

}
