% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateHubbell.R
\name{simulateHubbell}
\alias{simulateHubbell}
\title{Hubbell's neutral model simulation}
\usage{
simulateHubbell(
  n_species,
  M,
  I = 1000,
  d = 10,
  m = 0.02,
  tskip = 0,
  tend,
  norm = FALSE
)
}
\arguments{
\item{n_species}{integer amount of different species initially
in the local community}

\item{M}{integer amount of different species in the metacommunity,
including those of the local community}

\item{I}{integer value of fixed amount of individuals in the local community
(default: \code{I = 1000})}

\item{d}{integer value of fixed amount of deaths of local community
individuals in each generation (default: \code{d = 10})}

\item{m}{numeric immigration rate: the probability that a death in the local
community is replaced by a migrant of the metacommunity rather than by
the birth of a local community member (default: \code{m = 0.02})}

\item{tskip}{integer number of generations that should not be included
in the outputted species abundance matrix. (default: \code{tskip = 0})}

\item{tend}{integer number of simulations to be simulated}

\item{norm}{logical scalar choosing whether the time series should be
returned with the abundances as proportions (\code{norm = TRUE}) or
the raw counts (default: \code{norm = FALSE})}
}
\value{
\code{simulateHubbell} returns an abundance matrix with
species abundance as rows and time points as columns
}
\description{
Neutral species abundances simulation according to the Hubbell model.
}
\examples{
ExampleHubbell <- simulateHubbell(n_species = 8, M = 10, I = 1000, d = 50,
                                                        m = 0.02, tend = 100)

}
\references{
Rosindell, James et al. "The unified neutral theory of
biodiversity and biogeography at age ten." Trends in ecology & evolution
vol. 26,7 (2011).
}
\seealso{
\code{\link[miaSim:convertToSE]{convertToSE}}
}
