% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateHubbellRates.R
\name{simulateHubbellRates}
\alias{simulateHubbellRates}
\title{Hubbell's neutral model simulation applied to time series}
\usage{
simulateHubbellRates(
  community.initial,
  migration.p = 0.01,
  metacommunity.p = NULL,
  k.events = 1,
  growth.rates = NULL,
  norm = FALSE,
  t.end = 1000,
  ...
)
}
\arguments{
\item{community.initial}{Numeric: a vector of integers, containing species
counts greater or equal to zero.}

\item{migration.p}{Numeric: immigration possibility. It defines the
probability of replacement in the community. The value can be equal
be between 0 and 1. The sum of probability of migration and the probability
birth has to be 1.}

\item{metacommunity.p}{Numeric: the probability to find a species in the
metacommunity is the weighted average of probabilities in communities.}

\item{k.events}{Integer:}

\item{growth.rates}{Numeric: the rate of the change in the community size}

\item{norm}{Logical: whether the time series should be returned with
the abundances as proportions (\code{norm = TRUE}) or
the raw counts (default: \code{norm = FALSE})}

\item{t.end}{Numeric: simulation end time (default: \code{t.end = 1000})}

\item{...}{additional parameters including 't.start', 't.step', and 't.store'}
}
\value{
a list object containing the abundance matrix of each species,
number of individuals of each species before and after the simulation and the
metacommunity probability
}
\description{
Neutral species abundances simulation according to the Hubbell model.This
model shows that losses in society can be replaced either by the birth of
individuals or by immigration depending on their probabilities.
The specific time between the events of birth or migration is calculated and
time effect is considered to determine the next event.
}
\examples{
com_1 <- c(0, 5, 10)
com_2 <- c(20, 40, 60)
community.initial <- t(array(c(com_1, com_2)))
HubbellRates <- simulateHubbellRates(community.initial, migration.p = 0.01,
             metacommunity.p = NULL, k.events = 1, growth.rates = NULL,
             norm = FALSE, t.end=1000)

}
\references{
Rosindell, James et al. "The unified neutral theory of
biodiversity and biogeography at age ten." Trends in ecology & evolution
vol. 26,7 (2011).
}
