% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateHubbellRates.R
\name{simulateHubbellRates}
\alias{simulateHubbellRates}
\title{Hubbell's neutral model simulation applied to time series}
\usage{
simulateHubbellRates(
  community_initial,
  migration_p = 0.1,
  metacommunity_p = NULL,
  k_events = 1,
  growth_rates = NULL,
  norm = FALSE,
  t.end = 1000,
  list = TRUE,
  ...
)
}
\arguments{
\item{community_initial}{numeric value a vector of integers,
containing species counts greater or equal to zero.}

\item{migration_p}{numeric immigration possibility. It defines the
probability of migration when replacement is needed in the community.
The value can be between 0 and 1. The sum of the probability of
migration and the probability birth must be 1.}

\item{metacommunity_p}{numeric value the probability of a species being
found in the metacommunity.}

\item{k_events}{integer number of steps performed at a time point.
It can be equal or more than 1. Bigger k_events increases speed while
decreasing precision.}

\item{growth_rates}{numeric rate of change in community size.}

\item{norm}{logical scalar choosing whether the time series should be
returned with the abundances as proportions (\code{norm = TRUE}) or
the raw counts (default: \code{norm = FALSE})}

\item{t.end}{numeric value of simulation end time
(default: \code{t.end = 1000})}

\item{list}{logical scalar deciding whether output is a list object or not
(default: \code{norm = TRUE})}

\item{...}{additional parameters including 't.start', 't.step', and 't.store'}
}
\value{
a community abundance matrix or a list object that contains
growth rates, time points and metacommunity probabilities
}
\description{
Neutral species abundances simulation according to the Hubbell model. This
model shows that losses in society can be replaced either by the birth of
individuals or by immigration depending on their probabilities.
The specific time between the events of birth or migration is calculated and
time effect is considered to determine the next event.
}
\examples{
ExampleHubbellRates <- simulateHubbellRates(community_initial = c(0,5,10),
             migration_p = 0.01, metacommunity_p = NULL, k_events = 1,
             growth_rates = NULL, norm = FALSE, t.end=1000)

}
\references{
Rosindell, James et al. "The unified neutral theory of
biodiversity and biogeography at age ten." Trends in ecology & evolution
vol. 26,7 (2011).
}
\seealso{
\code{\link[miaSim:convertToSE]{convertToSE}}
}
