% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateRicker.R
\name{simulateRicker}
\alias{simulateRicker}
\title{Generate time series with the Ricker model}
\usage{
simulateRicker(
  n.species,
  A,
  x = runif(n.species),
  K = runif(n.species),
  sigma = 0.05,
  explosion.bound = 10^8,
  tskip = 0,
  tend,
  norm = FALSE
)
}
\arguments{
\item{n.species}{Integer: number of species}

\item{A}{interaction matrix}

\item{x}{Numeric: initial abundances}

\item{K}{Numeric: carrying capacities}

\item{sigma}{noise level, if set to a non-positive value, no noise is added
(default: \code{sigma = 0.05})}

\item{explosion.bound}{boundary for explosion
(default: \code{explosion.bound = 10^8})}

\item{tskip}{Integer: number of generations that should
not be included in the outputted species abundance matrix.
(default: \code{tskip = 0})}

\item{tend}{Integer: number of simulations to be simulated}

\item{norm}{Logical scalar: returns normalised abundances (proportions
in each generation) (default: \code{norm = FALSE})}
}
\value{
\code{simulateRicker} returns an abundance matrix with species abundance
as rows and time points as columns
}
\description{
The Ricker model is a discrete version of the generalized
Lotka-Volterra model and is implemented here as proposed by Fisher and Mehta
in PLoS ONE 2014.
}
\examples{
A <- powerlawA(10, alpha = 1.01)
ExampleRicker <- simulateRicker(n.species=10,A,tend=100)

}
\references{
Fisher & Mehta (2014). Identifying Keystone Species in the Human
Gut Microbiome from Metagenomic Timeseries using Sparse Linear Regression.
PLoS One 9:e102451
}
