% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateSOI.R
\name{simulateSOI}
\alias{simulateSOI}
\title{Self-Organised Instability model (SOI) simulation}
\usage{
simulateSOI(n.species, I, A, k = 5, com = NULL, tend, norm = FALSE)
}
\arguments{
\item{n.species}{Integer: number of species}

\item{I}{Integer: community size, number of available sites (individuals)}

\item{A}{interaction matrix}

\item{k}{Integer: the number of transition events that are allowed to take
place during one leap. (default: \code{k = 5}). Higher values reduce runtime,
but also accuracy of the simulation.}

\item{com}{a vector: initial community abundances
If (default: \code{com = NULL}), based on migration rates}

\item{tend}{Integer: number of timepoints to be returned in the time series
(number of generations)}

\item{norm}{Logical: indicates whether the time series should be returned
with the abundances as proportions (\code{norm = TRUE})
or the raw counts (default: \code{norm = FALSE})}
}
\value{
\linkS4class{SummarizedExperiment} object containing abundance matrix
consisting of species abundance as rows and time points as columns
}
\description{
Generate time-series with The Self-Organised Instability (SOI) model.
Implements a K-leap method for accelerating stochastic simulation.
}
\examples{
A <- miaSim::powerlawA(10, alpha = 1.2)

ExampleSOI <- simulateSOI(n.species = 10, I = 1000, A, k=5, com = NULL,
                                            tend = 150, norm = TRUE)

}
