% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulateSOI.R
\name{simulateSOI}
\alias{simulateSOI}
\title{Self-Organised Instability Model Community Simulation}
\usage{
simulateSOI(N, I, A, k, com, tend, norm = FALSE)
}
\arguments{
\item{N}{Integer: number of species}

\item{I}{Integer: community size, number of available sites (individuals)}

\item{A}{a interaction matrix of dimension NxN}

\item{k}{Integer: the number of transition events that are allowed to take
place during one leap. By default set to 5. Higher values reduce runtime,
but also accuracy of the simulation.}

\item{com}{a vector inditicating initial community abundances.
If NULL (default), based on migration rates.}

\item{tend}{Integer : number of timepoints to be returned in the time series
(nr of generations)}

\item{norm}{logical to indicate whether the time series should be returned
with the abundances as proportions (norm = TRUE)
or the raw counts (norm = FALSE, default)}
}
\value{
\linkS4class{SummarizedExperiment} object with abundance matrix
consisting of species abundance as rows and time points as columns
}
\description{
Generate time-series with the self-organised instability (SOI) model.
Implements a K-leap method for accelerating stochastic simulation.
}
\examples{
simulateSOI(N = 10, I = 1000, A = powerlawA(n = 10, alpha = 1.2),
                    k=5, com = NULL, tend = 150, norm = TRUE)
}
