% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateStochasticLogistic.R
\docType{methods}
\name{simulateStochasticLogistic}
\alias{simulateStochasticLogistic}
\alias{simulateStochasticLogistic-numeric}
\alias{simulateStochasticLogistic,numeric-method}
\title{Stochastic Logistic simulation}
\usage{
simulateStochasticLogistic(
  n.species,
  b = runif(n = n.species, min = 0.1, max = 0.2),
  k = runif(n = n.species, min = 1000, max = 2000),
  dr = runif(n = n.species, min = 5e-04, max = 0.0025),
  x = runif(n = n.species, min = 0.1, max = 10),
  t.start = 0,
  t.end = 2000,
  t.step = 0.01,
  t.store = 1000,
  partial = TRUE,
  stochastic = TRUE
)
}
\arguments{
\item{n.species}{Integer: number of species}

\item{b}{Numeric: growth rates
(default: \code{b = runif(n = n.species, min = 0.1, max = 0.2)})}

\item{k}{Numeric: carrying capacities
(default: \code{k = runif(n = n.species, min = 1000, max = 2000)})}

\item{dr}{Numeric: death rates
(default: \code{dr = runif(n = n.species, min = 0.0005, max = 0.0025)})}

\item{x}{Numeric: initial abundances
(default: \code{x = runif(n = n.species, min = 0.1, max = 10)})}

\item{t.start}{Numeric scalar indicating the initial time of the simulation.
(default: \code{t.start = 0})}

\item{t.end}{Numeric scalar indicating the final time of the dimulation
(default: \code{t.end = 2000})}

\item{t.step}{Numeric scalar indicating the interval between simulation steps
(default: \code{t.step = 0.01})}

\item{t.store}{Integer scalar indicating the number of evenly distributed
time points to keep
(default: \code{t.store = 1000})}

\item{partial}{Logical: whether to export only the stored time points
(default: \code{partial = TRUE})}

\item{stochastic}{Logical: whether the logistic model should be stochastic
(controlled by multiplying the growth rate by a random number)
(default: \code{stochastic = TRUE})}
}
\value{
\code{simulateStochasticLogistic} returns a
\linkS4class{SummarizedExperiment} class object containing matrix with
species abundance as rows and time points as columns
}
\description{
Simulates time series with the (stochastic) logistic model and forms a
\linkS4class{SummarizedExperiment} object.
}
\details{

}
\examples{
## ATTENTION: Don't set a large value to t.step, otherwise the computer won't
#give a correct solution to the logistic ODE(ordinary differential equation).
#Keeping t.step under 0.05 or 0.01 is a good practice.

#while (!exists("ExampleLogistic"))
ExampleLogistic <- simulateStochasticLogistic(n.species = 5)
#plot the calculated points
matplot(t(assays(ExampleLogistic)[["counts"]]), type = "l")

#calculation by setting initial parameters explicitly
ExampleLogistic <- simulateStochasticLogistic(
n.species = 2,
b <- c(0.2, 0.1), k <- c(1000, 2000), dr <- c(0.001, 0.0015), x <- c(3, 0.1),
t.start = 0, t.end = 1500, t.step = 0.01,
t.store = 1200, partial = FALSE, stochastic = FALSE)

}
