% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{methods}
\name{tDyn}
\alias{tDyn}
\alias{tDyn-numeric}
\alias{tDyn,numeric-method}
\title{Generate simulation times and the indices
of time points to return in sumulation functions.}
\usage{
tDyn(t.start = 0, t.end = 1000, t.step = 0.1, t.store = 100)
}
\arguments{
\item{t.start}{Numeric scalar indicating the initial time of the simulation.
(default: \code{t.start = 0})}

\item{t.end}{Numeric scalar indicating the final time of the dimulation
(default: \code{t.end = 1000})}

\item{t.step}{Numeric scalar indicating the interval between simulation steps
(default: \code{t.step = 0.1})}

\item{t.store}{Integer scalar indicating the number of evenly distributed
time points to keep (default: \code{t.store = 100})}
}
\value{
list containing simulation times (t.sys) and the indices to keep.
}
\description{
Generate simulation times and the indices
of time points to return in sumulation functions.
}
\examples{
Time <- tDyn(t.start = 0, t.end = 100, t.step = 0.5, t.store = 100)
DefaultTime <- tDyn(t.start = 0)

}
\keyword{internal}
