% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tDyn}
\alias{tDyn}
\title{Generate simulation times and the indices
of time points to return in sumulation functions.}
\usage{
tDyn(t.start = 0, t.end = 1000, t.step = 0.1, t.keep = 100)
}
\arguments{
\item{t.start}{Numeric scalar. Initial time of the simulation, default 0.}

\item{t.end}{Numeric scalar. Final time of the simulation, default 1000.}

\item{t.step}{Numeric scalar. Interval between simulation steps, default 0.1.}

\item{t.keep}{Integer scalar. Number of evenly distributed time points to keep, default 100.}
}
\value{
list containing simulation times (t.sys) and the indices to keep.\
}
\description{
Generate simulation times and the indices
of time points to return in sumulation functions.
}
\keyword{internal}
